refactor: team-dashboard 컴포넌트 분리
